[wip] Finish azure snapshots proof-of-concept
This patch makes azure snapshot backup and restore functional with
barman cloud. It highlights the areas where we're going to need to
re-think the existing abstraction a little - mostly because things
which seemed generic when implementing GCP (like the device a disk
is attached at being consistent, and zones being used in grouping
resources) are actually specific to GCP, or at least they don't
really work the same way in Azure.

The big two things are:

- Availability zones are just a detail and are not used to locate
  resources. Resource groups serve that purpose instead, so currently
  we pass a resource group as the zone paramater in the cli. This is
  obviously silly.
- Device names are ephemeral - the device name an attached disk ends
  up with is not guaranteed. The deterministic property for disk
  attachment is the LUN - this will always be the same. That means
  that there's no real point in storing device names with Azure since
  they are not useful, and it also means that the LUN must be resolved
  to a device name in order for the logic which fetches and checks
  mount points and mount options to work.

These aren't blockers but they will require some careful thought and
a little refactoring, i.e. these two issues are where The Real Work
for this change will be required.

Example backup command:

    barman-cloud-backup --cloud-provider=azure-blob-storage \
        https://barmandev1.blob.core.windows.net/edb-backup-mike-testing/prefix \
        test-server \
        --snapshot-instance primary-vm \
        --snapshot-zone=barman-av-snapshots-uksouth-rg \
        --snapshot-disk=barman-av-snapshots-uksouth-primary--dev-sdc \
        --snapshot-disk=barman-av-snapshots-uksouth-primary--dev-sdd \
        --snapshot-azure-subscription-id=ac9903a5-bfd9-4576-a9cc-a463f5e41079