[wip] AWS EBS snapshots part 1
Does not make real snapshot API call or delete snapshots yet, or
figure out devices for recovery.

Also assumes kernel does not rename devices but the kernel can in
fact rename devices. So there's that too.

And endpoint_url needs to be passed in from barman cloud CLI or
barman config in order to initialise the snapshot interface.