[gcp-refactor] Remove knowledge of zone from barman code
Zone is specific to GCP so the barman code should not have knowledge
of it. Because there are also other arguments which are specific to
other cloud providers we remove all handling of zone from barman and
rely on it being either passed in a dict of `provider_args` or being
available in the config/args. In all cases, the only code which needs
to care is the factory function which creates the interface.

This also requires moving the zone to be a property of the interface
instead of something the backup/recovery executors care about. This
means once the interface has been created, everything must happen
in that zone.

For backup it means the disks and instance must be in the same zone.
At recovery time, the disks and instance must again be in the same
zone but this need not tbe the zone the backup instance/disks were
in. A different zone can be specified.