[wip] Refactor mapping of snapshots, volumes and mounts
Refactors the way snapshots and volumes are mapped to mount points
so that the main backup/recovery code paths do not need to concern
themselves with the provider-specific differences in how this is
achieved.

The motivation for this commit was the fact that we need to be able
to resolve storage volumes as known by the cloud provider to attached
volumes on the backup or recovery VM. On GCP this could be done by
using the device name available in the instance metadata, however
with other cloud providers the process is slightly different:

- On AWS the device name can also be used however the actual device
  name on the instance *may* be different to the one provided in the
  instance metadata returned by the API.
- On Azure, the device name is not provided in the instance metadata at
  all because it is assigned at boot time and is not a permanent
  property of the instance. The constant attribute between attached
  volumes and mounted volumes is the LUN - this is included in the
  instance metadata and can be resolved to the mounted volume on the
  instance.

Because both the metadata used to resolve the attached volume *and* the
resolution process itself can vary, it is necessary to abstract it
so that calling code can just deal with the data it needs from the
resolved volume (the mount point and mount options) and not worry about
the details of the process.

This resolution process happens at:

- Check time, when verifying the named disks for the snapshot backup
  are attached to the instance.
- Backup time, when recording the mount point and mount options of each
  disk from which a snapshot was taken.
- Recovery time, when verifying a disk cloned from each snapshot is
  attached to the recovery instance at the expected mount point (and
  with the expected options).

This was achieved by abstracting device name behind a `device` property
in the SnapshotMetadata class. This was intended to allow the details
of determining the device to vary by cloud provider, however this alone
is not enough abstraction because the device itself is not necessarily
useful information and because of the previously mentioned differences
in how the mounted volume is resolved.

A new abstraction called VolumeMetadata is therefore added. This
represents the metadata for an attached and mounted volume. Unlike
SnapshotsMetadata, which was intended to represent information stored
in the `backup.info` file and also provide a provider-independent
abstraction over attachment metadata, VolumeMetadata is purely intended
to represent an attached volume at a moment in time so that Barman can
resolve volume names and snapshots to mounted volumes.

The device field is removed from SnapshotMetadata since it is no longer
needed for the purposes of resolving volumes to attached volumes. We
retain the device_name field in the GcpSnapshotMetadata because this is
useful information which, while not being *necessary* at recovery time,
might still be useful.